kernel: add compat for upstream new features
refer commit/files
https://github.com/backslashxx/KernelSU/commit/02120f0e49e82d9d368e6880e804619057cca912
https://github.com/backslashxx/KernelSU/blob/master/kernel/su_mount_ns.c#L144
https://github.com/rsuntk/KernelSU/blob/main/kernel/selinux/sepolicy.c
https://github.com/rsuntk/KernelSU/commit/228331be677f694d33941d7b6958d2e555549b29

Tested have passed in
5.4.300 sunstone kernel
4.9.117 cactus kernel

+++ kernel/su_mount_ns.c
@@ -125,7 +125,7 @@ try_setns:
         goto out;
     }
     struct path ns_path;
-    long ret = ns_get_path(&ns_path, pid1_task, &mntns_operations);
+    long ret = (long)ns_get_path(&ns_path, pid1_task, &mntns_operations);
     put_task_struct(pid1_task);
     if (ret) {
         pr_warn("failed get path for init mount namespace: %ld\n", ret);
